requirements.txt
Click
Flask
itsdangerous
Jinja2
MarkupSafe
Werkzeug
flask-httpauth
psycopg2
flask-sqlalchemy
Flask-Migrate
sqlalchemy-migrate
flask-restful
python-dotenv
pyjwt
flask-bcrypt
requests
